pyproject.toml
[project]
name = "hable-ya"
version = "0.1.0"
description = "On-device Spanish language acquisition voice agent"
requires-python = ">=3.12"
dependencies = [
    # API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=13.0",
    # Pipecat
    "pipecat-ai[silero,daily]>=0.0.103",
    # STT / TTS
    "faster-whisper>=1.0.0",
    "piper-tts>=1.2.0",
    # LLM client
    "openai>=1.50.0",               # OpenAI-compatible llama.cpp endpoint
    # DB
    "asyncpg>=0.30.0",              # async Postgres driver
    "alembic>=1.13.0",              # migration tool
    "sqlalchemy[asyncio]>=2.0.0",   # required by alembic; used only in env.py
    # Config
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    # Utilities
    "langdetect>=1.0.9",
    "tenacity>=9.0.0",              # retries for tool calls
    "jupyterlab>=4.5.6",
    # NLP — Spanish lemmatization for runtime vocabulary tracking (#029) and
    # recast scoring (eval). The es_core_news_sm model is a manual setup step
    # (`python -m spacy download es_core_news_sm`); VocabularyRepo degrades
    # gracefully if the model isn't loadable.
    "spacy>=3.7",
]

[project.optional-dependencies]
eval = [
    "anthropic>=0.40.0",           # Opus for synthetic learner + judge
    "rich>=13.0.0",                # pretty eval output
    "pandas>=2.0.0",               # results aggregation
]
finetune = [
    "anthropic>=0.40.0",           # Opus for skeleton generation
    "datasets>=3.0.0",             # HuggingFace datasets
    "unsloth",                     # SFT / DPO training (replaces trl + peft)
    "transformers>=4.47.0",
    "torch>=2.5.0",
    "huggingface_hub>=0.25.0",     # model downloads
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "rich>=13.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "rich>=13.0.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.lint.per-file-ignores]
# themes.py is hand-authored scenario data — wrapping target_structures
# lists one-item-per-line quadruples the diff for zero readability gain.
"hable_ya/learner/themes.py" = ["E501"]
# Alembic migrations are SQL-heavy; wrapping mid-statement hurts readability
# more than it helps.
"hable_ya/db/alembic/versions/*.py" = ["E501"]
# Prompt templates and judge rubrics are hand-authored prose (English/Spanish
# example sentences, CLI usage examples in docstrings). Wrapping mid-sentence
# damages readability of the prompt itself.
"scripts/fixtures/prompts/*.py" = ["E501"]
"eval/judge_recasts.py" = ["E501"]
"eval/run_eval.py" = ["E501"]
"scripts/benchmark_concurrency.py" = ["E501"]
"finetune/format.py" = ["E501"]
"finetune/generate.py" = ["E501"]
"notebooks/*.ipynb" = ["E501"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
exclude = [
    # Auto-generated by Unsloth during training — not our code, not worth typing.
    "notebooks/unsloth_compiled_cache/",
    "unsloth_compiled_cache/",
]

[[tool.mypy.overrides]]
# Tests, utility scripts, eval harnesses, and fine-tune glue are best-effort:
# we still catch import errors and real type mismatches via default checks,
# but don't require full strict annotations. Core library code
# (hable_ya/, api/) remains under strict typing.
module = [
    "tests.*",
    "scripts.*",
    "eval.*",
    "finetune.*",
]
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_decorators = false
disallow_untyped_calls = false
disallow_any_generics = false
warn_return_any = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]

[tool.hatch.build.targets.wheel]
packages = ["hable_ya", "api"]
